@@ -37,8 +37,8 @@ dependencies:
   shared_preferences: ^2.2.2
   image_picker: ^1.0.7
   http: ^1.2.0
-  connectivity_plus: ^4.0.0
-  webview_flutter: ^4.11.0
+  connectivity_plus: ^6.1.4
+  webview_flutter: ^4.13.0
   url_launcher: ^6.1.5  # Add this line
   build_runner: ^2.4.0
   build_web_compilers: ^4.0.4
@@ -54,6 +54,8 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^5.0.0
+  mockito: ^5.4.4
+  build_runner: ^2.4.0 # Ensure it's also here if needed for test generation commands
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -66,8 +68,8 @@ flutter:
   # the material Icons class.
   uses-material-design: true
   
-  assets:
-   # - assets
+  # assets:
+  #  # - assets
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
