@@ -3,7 +3,7 @@ import 'home_screen.dart';
 //import 'package:google_mobile_ads/google_mobile_ads.dart';
 
 void main() async {
- // WidgetsFlutterBinding.ensureInitialized();
+  // WidgetsFlutterBinding.ensureInitialized();
   //await MobileAds.instance.initialize();
 
   runApp(MyApp());
@@ -25,15 +25,17 @@ class MyApp extends StatelessWidget {
           secondary: const Color(0xFFF06292),
           tertiary: const Color(0xFFAD1457),
           surface: const Color(0xFFFCE4EC),
+          background: const Color(0xFFFCE4EC), // Added background color
         ),
         textTheme: const TextTheme(
           displayLarge: TextStyle(
             fontSize: 42,
             fontWeight: FontWeight.bold,
-            fontFamily: 'Roboto',
+            fontFamily: 'Roboto', // Consider if Roboto is consistently used/imported
           ),
           titleLarge: TextStyle(fontSize: 24, fontWeight: FontWeight.w600),
           bodyMedium: TextStyle(fontSize: 16),
+          labelLarge: TextStyle(fontSize: 18, fontWeight: FontWeight.bold), // Added for buttons
         ),
         scaffoldBackgroundColor: const Color(0xFFFCE4EC),
         appBarTheme: const AppBarTheme(
